requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.27
pymysql==1.1.1
python-multipart==0.0.9
jinja2==3.1.6
aiofiles==23.2.1 